% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{genotypeMatrix}
\alias{genotypeMatrix}
\title{imports genotypeMatrix}
\usage{
genotypeMatrix(
  fileName,
  N,
  P,
  col_skip_pos,
  col_skip,
  keepbytes,
  keepoffset,
  fillmissing
)
}
\arguments{
\item{fileName}{location of plink binary file}

\item{N}{number of subjects}

\item{P}{number of SNPs}

\item{col_skip_pos}{which variants to skip}

\item{col_skip}{which variants to skip}

\item{keepbytes}{which bytes to keep}

\item{keepoffset}{the offset}

\item{fillmissing}{whether to fillin missing cells}
}
\value{
RcppArmadillo genotype matrix
}
\description{
imports genotypeMatrix
}
