% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{normalize2}
\alias{normalize2}
\title{normalize genotype matrix}
\usage{
normalize2(genotypes)
}
\arguments{
\item{genotypes}{RcppArmadillo genotype matrix}
}
\value{
vector of standard deviations
}
\description{
normalize genotype matrix
}
