% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudoAicBic.R
\name{pseudoAicBic}
\alias{pseudoAicBic}
\title{Function to obtain pseudo-AIC and pseudo-BIC estimates given summary statistics and a reference panel}
\usage{
pseudoAicBic(
  penalizedBetas,
  betas,
  ses,
  N,
  refPanel,
  sigSqReg = 0.2,
  sseReg = 0.2,
  sigSqInd = NULL,
  allele1 = NULL,
  allele2 = NULL,
  standardized = TRUE,
  extract = NULL,
  ldBlocks = NULL
)
}
\arguments{
\item{penalizedBetas}{Polygenic risk score effect size estimates. These are the estimates for which we want to approximate AIC and BIC. May be a matrix.}

\item{betas}{Univariate effect size estimates.}

\item{ses}{Standard errors of the univariate effect size estimates.}

\item{N}{Sample sizes corresponding to the univariate effect size estimates. Can be a constant or a vector.}

\item{refPanel}{Stem of PLINK binary file}

\item{sigSqReg}{The matrix regularization parameter for the estimation of the residual variance. Default is .2.}

\item{sseReg}{The matrix regularization parameter for the estimation of the SSE. Default is .2.}

\item{sigSqInd}{Index of SNPs to be used for estimation of residual variance. If null, default is to pick the n/5 SNPs with largest univariate effect sizes, where n is the sample size of the reference panel.}

\item{allele1}{Vector of effect alleles for the betas and penalizedBetas. Corresponds to the fifth column of a PLINK .bim file.}

\item{allele2}{Vector of reference alleles for the betas and penalizedBetas. Corresponds to the sixth column of a PLINK .bim file.}

\item{standardized}{Set to true if the coefficient estimates for penalizedBetas are standardized. Note that elastSum and tlpSum output standardized estimates.}

\item{extract}{Vector of the indices of SNPs to keep. If null, will keep all SNPs.}

\item{ldBlocks}{Location of file specifying independent LD Blocks to be used. File should be in the BED file format. If null, estimation is done by chromosome.}
}
\description{
Function to obtain pseudo-AIC and pseudo-BIC estimates given summary statistics and a reference panel
}
\details{
Function requires univariate effect size estimates, standard errors, and sample sizes, candidate polygenic risk score estimates, and a reference panel.
}
