% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quasicors.R
\name{quasicors}
\alias{quasicors}
\title{Function to obtain quasi-correlation estimates given}
\usage{
quasicors(testBetas, penalizedBetas, refPanel, allele1 = NULL,
  allele2 = NULL, trainBetas = NULL, trainSes = NULL,
  standardized = TRUE)
}
\arguments{
\item{testBetas}{Univariate effect sizes for the out-of-sample data}

\item{penalizedBetas}{The polygenic risk scores to be tested on the out-of-sample data. May be a matrix.}

\item{refPanel}{PLINK stem for binary file to be used as the reference panel}

\item{allele1}{Vector of effect alleles for the penalizedBetas. Corresponds to the fifth column of a PLINK .bim file.}

\item{allele2}{Vector of reference alleles for the penalizedBetas. Corresponds to the sixth column of a PLINK .bim file.}

\item{trainBetas}{Vector of univariate effect sizes for the data used to estimate the polygenic risk score. Used to estimate variance of training phenotype. Unneccessary if standardized = FALSE.}

\item{standardized}{Set to true if the coefficient estimates for penalizedBetas are standardized. Note that elastSum and tlpSum output standardized estimates.}

\item{trainBetas}{Vector of univariate standard errors for the data used to estimate the polygenic risk score. Used to estimate variance of training phenotype. Unneccessary if standardized = FALSE.}
}
\value{
a list with the following
\item{quasicorrelations} Quasi-correlation values for each of the input polygenic risk scores.
\item{bxy} The estimate of the dot product of the univariate out-of-sample effect size estimates and the candidate polygenic risk score: b'x'y.
\item{bxxb} The estimate of the variance of the predicted phenotype for the out of sample data: b'x'xb.
}
\details{
Function requires univariate effect size estimates, standard errors, and sample sizes, candidate polygenic risk score estimates, and a reference panel.
}
