% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tlpSum.R
\name{tlpSum}
\alias{tlpSum}
\title{Function to obtain TLP effect size estimates given summary statistics and a reference panel}
\usage{
tlpSum(cors, bfile, lambdas, taus, s = 0.5, thr = 1e-04, init = NULL,
  maxIter = 1000, extract = NULL, ldBlocks = NULL)
}
\arguments{
\item{cors}{A vector of correlations, representing the univariate SNP-phenotype associations from the summary statistic data.}

\item{bfile}{The stem of a PLINK binary file to be used as a reference panel.}

\item{lambdas}{Vector of values of lambda.}

\item{taus}{Vector of values of tau.}

\item{s}{Vector of values for s.}

\item{thr}{Convergence threshold.}

\item{init}{Initial values for \eqn{\beta}.}

\item{maxIter}{Maximum number of iterations.}

\item{extract}{Vector of the indices of SNPs to keep. If null, will keep all SNPs.}

\item{ldBlocks}{Location of file specifying independent LD Blocks to be used. File should be in the BED file format. If null, estimation is done by chromosome.}
}
\value{
a list with the following
\item{lambdas} Vector of lambda values corresponding to columns of output files 'beta' and 'pred'
\item{taus} Vector of tau values corresponding to columns of output files 'beta' and 'pred'
\item{s} Vector of s values corresponding to columns of output files 'beta' and 'pred'
\item{beta} Matrix of estimated coefficients.
\item{converged} Matrix of convergence indicators with following format: column 1 is lambda value, column 2 is tau value, column 3 is s value, column 4 is 1 if there was convergence.
\item{pred} Matrix of predicted phenotypes.
}
\details{
Function finds the minimum of the TLP objective function
}
