% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudoAicBic.R
\name{pseudoAicBic}
\alias{pseudoAicBic}
\title{Function to obtain pseudo-AIC and pseudo-BIC estimates given summary statistics and a reference panel}
\usage{
pseudoAicBic(penalizedBetas, betas, ses, N, refPanel, sigSqReg = 0.2,
  sseReg = 0.1, sigSqInd = NULL)
}
\arguments{
\item{penalizedBetas}{Polygenic risk score effect size estimates. These are the estimates for which we want to approximate AIC and BIC. May be a matrix.}

\item{betas}{Univariate effect size estimates.}

\item{ses}{Standard errors of the univariate effect size estimates.}

\item{N}{Sample sizes corresponding to the univariate effect size estimates. Can be a constant.}

\item{refPanel}{Stem of PLINK binary file}

\item{sigSqReg}{The matrix regularization parameter for the estimation of the residual variance. Default is .2.}

\item{sseReg}{The matrix regularization parameter for the estimation of the SSE. Default is .1.}

\item{sigSqInd}{Index of SNPs to be used for estimation of residual variance. If null, default is to pick the n/5 SNPs with largest univariate effect sizes, where n is the sample size of the reference panel.}
}
\value{
a list with the following
\item{aic} Estimates of AIC for the candidate polygenic risk scores.
\item{bic} Estimates of BIC for the candidate polygenic risk scores.
\item{SSE} Estimates of SSE for the candidate polygenic risk scores.
\item{q} Number of active parameters for the candidate polygenic risk scores.
\item{bxxb} Estimates of the b'x'xb quantity for the candidate polygenic risk scores
\item{bxy} Estimates of b'x'y for the candidate polygenic risk scores.
\item{sigSqTilde} Estimate of the residual variance.
}
\details{
Function requires univariate effect size estimates, standard errors, and sample sizes, candidate polygenic risk score estimates, and a reference panel.
}
