% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quasicors.R
\name{quasicors}
\alias{quasicors}
\title{Function to obtain quasi-correlation estimates given}
\usage{
quasicors(
  testBetas,
  testSes,
  N,
  penalizedBetas,
  refPanel,
  allele1 = NULL,
  allele2 = NULL,
  standardized = TRUE,
  trainVar = NULL,
  allele1.test = NULL,
  allele2.test = NULL
)
}
\arguments{
\item{testBetas}{Univariate effect sizes for the out-of-sample data}

\item{testSes}{Univariate standard errors for the effect sizes of the out-of-sample data}

\item{N}{Sample sizes corresponding to the testBetas. Can be a constant or a vector.}

\item{penalizedBetas}{The polygenic risk scores to be tested on the out-of-sample data. May be a matrix.}

\item{refPanel}{PLINK stem for binary file to be used as the reference panel}

\item{allele1}{Vector of effect alleles for the penalizedBetas. Corresponds to the fifth column of a PLINK .bim file.}

\item{allele2}{Vector of reference alleles for the penalizedBetas. Corresponds to the sixth column of a PLINK .bim file.}

\item{standardized}{Set to true if the coefficient estimates for penalizedBetas are standardized. Note that elastSum and tlpSum output standardized estimates.}

\item{trainVar}{Phenotypic variance for data used to estimate the penalizedBetas. Only necessary if standardized = TRUE}

\item{allele1.test}{Vector of effect alleles for the testBetas. Corresponds to the fifth column of a PLINK .bim file.}

\item{allele2.test}{Vector of reference alleles for the testBetas. Corresponds to the sixth column of a PLINK .bim file.}

\item{ldBlocks}{Location of file specifying independent LD Blocks to be used. File should be in the BED file format. If null, estimation is done by chromosome.}
}
\description{
Function to obtain quasi-correlation estimates given
}
\details{
Function requires univariate effect size estimates, standard errors, and sample sizes, candidate polygenic risk score estimates, and a reference panel.
}
